working on adding a custom img to the browser tab
added pvr-LOGO-01 that i made but for some reason its not working. need to go into manifest and see whats going on with that. also tweaked the font sizing in the navbar so its a little bigger, and got rid of the ability to drag certain components that would get in the way of mobile scrolling. maybe reintroduce it on just the header card