added another photocard to aboutMePage
wanted to change up the flow of the about me page and i tweaked some of the component margin spacing to even the blank space in the about me section also tried to add some more meaningful test on the aboutme card and the motivation card. i think its a step in the right direction once i finish setting up contact form ill be getting close to finishing.